[IMP] medical_financial_coverage_agreement: Improve tests
