+++ medical_financial_coverage_agreement/tests/__init__.py
@@ -2,4 +2,5 @@
 # Copyright 2017 Eficent Business and IT Consulting Services, S.L.
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
+from . import common
 from . import test_coverage_agreement
